// src/soumu/blocks/datatable/datatable.type.tsx
export interface SortableQueries {
  dataKey: string,
  type: 'asc' | 'desc';
}

/**
 * Sample data table query structure
 * ```js
 * {
 *   page: 1,
 *   size: 20,
 *   sort: {
 *     key: 'data_key',
 *     type: 'asc',
 *   },
 *   filter: {
 *     filter_key_1: 'any type',
 *     filter_key_2: 'any type',
 *     ...
 *   }
 * }
 * ```
 */
export interface DataTableQueries<F = any> {
  /**
   * Default: 1
   */
  page: number;

  /**
   * Default: from API setting.
   */
  size: number;

  /**
   * Default: null
   */
  sort?: SortableQueries;

  /**
   * Default: null
   */
  filter?: F;
}

export interface HeaderMiddleProps<D = any> {
  /**
   * `filter` is `DataTableQueries.filter`.
   */
  onSearch?: (filter: { [filterKey: string]: any }) => void;

  /**
   * Set initial filter value
   */
  setFilter?: (filter: { [filterKey: string]: any }) => void;

  setRef?: (name, ref: any) => void;

  /**
   * Data is defined by `DataTableProps.headerMiddleData`.
   */
  data?: D;
}

export interface HeaderLowerProps {
  columnsConfig: ColumnsConfig<any>,
}

export interface CellComponentProps<DataItem = any, CustomData = any> {
  dataItem: DataItem;
  dataKey?: string;
  customData?: CustomData;
  /**
   * Current data item index;
   */
  index: number;
}

export type TablePageSizes = { label: string, value: number }[];

export type ColumnGroups = { startColumn: number, endColumn: number, label: string, }[];

export interface TableConfig<DataItem> {
  /**
  * This property must be used when you want to hidden header.
  */
  hideHeader?: boolean;

  /**
   * This property must be used when you want to override middle header block.
   */
  headerMiddle?: ComponentType<HeaderMiddleProps>;

  /**
   * This property must be used when you want to override lower header block.
   * - null: hide headerLower.
   * - ComponentType: use this component to render headerLower.
   *
   * By default, it have:
   * - pagination
   * - column display setting
   */
  headerLower?: ComponentType<HeaderLowerProps>;

  /**
   * Default: [{ label: 20件, value: 20 }, { label: 50件, value: 50 }, { label: 2100件, value: 100 }]
   * @deprecated Page size options will be getting from API.
   */
  pageSizes?: TablePageSizes;

  /**
   * _Weird but work_
   * Note: fixed column will not be able to setting as hidden.
   *
   * Default: 1
   */
  fixedColumnNumber?: number;

  /**
   * This property must be used when you don't want any fixed column
   */
  disableFixedColumn?: boolean;

  /**
   * TODO: support default value + rename.
   */
  rowSizes: number[];

  /**
   * The unique property name from `DataItem`. Usually you will use something like `id`.
   *
   * "For goodness, please give me the truth - your real id" - someone cried.
   */
  idProp: keyof DataItem & string;

  columnGroups?: ColumnGroups;

  screenCd?: string;
  tableId?: number;

  /**
   * Callback to custom condition to display radio input.
   */
  showRadio?: (dataItem: DataItem) => boolean;

  /**
    * Callback to custom condition to display checkbox input.
  */
  showCheckBox?: (dataItem: DataItem) => boolean;
}

export interface ColumnConfig<DataItem = any> {
  /**
   * Column name/label.
   */
  label: string;

  /**
   * Value must be a property key of `DataItem`.
   */
  dataKey?: keyof DataItem & string;

  /**
   * Whether this column must be displayed.
   *  _(User can use "setting display column" feature to change this value)._
   *
   * Default: true
   */
  isVisible?: boolean;

  /**
   * Define whether this column can be setting as hidden.
   * IMPORTANT: There is new logic, fixed column will not be able to setting as hidden.
   *
   * Default: false(user can setting this column as hidden column).
   * @see TableConfig.fixedColumnNumber
   */
  disabled?: boolean;

  /**
   * Specify component that will be used to display cell content.
   * By defaut, it display plain text, so there is no need to use this property
   * if you don't want to custom cell content.
   */
  component?: ComponentType<CellComponentProps<DataItem>>;

  /**
   * Custom data will be passed to component(`ColumnConfig.component`) as a property named `customData`.
   */
  customData?: any;

  isSorted?: any;

  align?: 'inherit' | 'left' | 'center' | 'right' | 'justify';

  width: string;
}

export type ColumnsConfig<DataItem = any> = ColumnConfig<DataItem>[];